Exclude code and assertion from source needed only for tests
